Create TwitterData service and refactor
