+++ sm-analytic/ClientApp/src/app/features/dashboard/dashboard.module.ts
@@ -18,6 +18,8 @@ import { ApiService } from '../../shared/services/api.service';
 import { ChartsModule } from 'ng2-charts';
 import { LineAreaChartComponent } from './components/charts/line-area-chart/line-area-chart.component';
 import { EngagementService } from 'app/shared/services/engagement.service';
+import { TwitterDataService } from 'app/shared/services/twitter-data.service';
+
 
 @NgModule({
   imports: [
@@ -40,7 +42,8 @@ import { EngagementService } from 'app/shared/services/engagement.service';
   ],
   providers: [
     ApiService,
-    EngagementService
+    EngagementService,
+    TwitterDataService
   ],
   exports: [
     DashboardComponent
